rename utils to common
