+++ autosubliminal/diskscanner.py
@@ -13,7 +13,7 @@ from enzyme.mkv import MKV
 
 import autosubliminal
 from autosubliminal import fileprocessor
-from autosubliminal.util.utils import add_event_message, add_notification_message, get_wanted_queue_lock, \
+from autosubliminal.util.common import add_event_message, add_notification_message, get_wanted_queue_lock, \
     release_wanted_queue_lock, release_wanted_queue_lock_on_exception, skip_movie, skip_show
 from autosubliminal.core.scheduler import ScheduledProcess
 from autosubliminal.db import WantedItems
