+++ autosubliminal/server/web/config.py
@@ -7,7 +7,7 @@ import autosubliminal
 from autosubliminal import config, notifiers
 from autosubliminal.server.web import redirect
 from autosubliminal.templates.page import PageTemplate
-from autosubliminal.util.utils import add_notification_message, get_boolean, mapping_string_to_dict
+from autosubliminal.util.common import add_notification_message, get_boolean, mapping_string_to_dict
 
 
 class Config(object):
