+++ autosubliminal/subchecker.py
@@ -26,7 +26,7 @@ from autosubliminal.postprocessor import PostProcessor
 from autosubliminal.providers import provider_cache
 from autosubliminal.providers.addic7ed import Addic7edSubtitle as Addic7edSubtitleRandomUserAgent
 from autosubliminal.subdownloader import SubDownloader
-from autosubliminal.util.utils import add_event_message, add_notification_message, get_wanted_queue_lock, \
+from autosubliminal.util.common import add_event_message, add_notification_message, get_wanted_queue_lock, \
     release_wanted_queue_lock, release_wanted_queue_lock_on_exception, set_rw_and_remove, wait_for_internet_connection
 
 log = logging.getLogger(__name__)
