+++ autosubliminal/versionchecker.py
@@ -17,7 +17,7 @@ import autosubliminal
 from autosubliminal import system
 from autosubliminal.core.enums import InstallType
 from autosubliminal.core.scheduler import ScheduledProcess
-from autosubliminal.util.utils import add_notification_message, connect_url, get_wanted_queue_lock, \
+from autosubliminal.util.common import add_notification_message, connect_url, get_wanted_queue_lock, \
     release_wanted_queue_lock, release_wanted_queue_lock_on_exception, wait_for_internet_connection
 from autosubliminal.version import RELEASE_VERSION
