+++ tests/util/cassettes/common/test_connect_url.yaml
@@ -15,13 +15,13 @@ interactions:
       age: ['0']
       connection: [keep-alive]
       content-length: ['0']
-      date: ['Thu, 05 Apr 2018 19:21:07 GMT']
+      date: ['Thu, 05 Apr 2018 20:04:42 GMT']
       location: ['https://raw.githubusercontent.com/h3llrais3r/Auto-Subliminal/master/autosubliminal/version.py']
       vary: [Accept-Encoding]
       via: [1.1 varnish]
       x-cache: [MISS]
       x-cache-hits: ['0']
-      x-fastly-request-id: [f3f32ca3840203c9fcaab87128a3054170ad8ab5]
+      x-fastly-request-id: [22614cef464473487d96e9fd3efde89ffaf710af]
       x-served-by: [cache-ams4137-AMS]
     status: {code: 301, message: Moved Permanently}
 - request:
@@ -47,22 +47,22 @@ interactions:
       content-length: ['79']
       content-security-policy: [default-src 'none'; style-src 'unsafe-inline'; sandbox]
       content-type: [text/plain; charset=utf-8]
-      date: ['Thu, 05 Apr 2018 19:21:07 GMT']
+      date: ['Thu, 05 Apr 2018 20:04:42 GMT']
       etag: ['"b26f08a580d46d31dca0bf89a56089b12b707d64"']
-      expires: ['Thu, 05 Apr 2018 19:26:07 GMT']
-      source-age: ['194']
+      expires: ['Thu, 05 Apr 2018 20:09:42 GMT']
+      source-age: ['56']
       strict-transport-security: [max-age=31536000]
       vary: ['Authorization,Accept-Encoding']
       via: [1.1 varnish]
       x-cache: [HIT]
       x-cache-hits: ['1']
       x-content-type-options: [nosniff]
-      x-fastly-request-id: [42124f4b910fefef5719b9ff0f2ac2a7016a51e7]
+      x-fastly-request-id: [3bf671fbcd7a0c846f5a1738c99d42ec9ce65bd4]
       x-frame-options: [deny]
       x-geo-block-list: ['']
-      x-github-request-id: ['D3B2:5CB2:2979:2BA1:5AC67661']
-      x-served-by: [cache-ams4148-AMS]
-      x-timer: ['S1522956068.665730,VS0,VE1']
+      x-github-request-id: ['D430:5CB3:2442E:25D38:5AC68121']
+      x-served-by: [cache-ams4437-AMS]
+      x-timer: ['S1522958683.583239,VS0,VE1']
       x-xss-protection: [1; mode=block]
     status: {code: 200, message: OK}
 version: 1
